requirements.txt
aniso8601==1.2.1
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
decorator==4.1.0
Flask==0.12.2
Flask-OAuth==0.12
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
geopy==1.11.0
giturlparse.py==0.0.5
httplib2==0.10.3
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
MarkupSafe==1.0
murl==0.5.1
oauth2==1.9.0.post1
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.2
PyMySQL==0.7.11
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.2
recordtype==1.1
requests==2.18.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.11
traitlets==4.3.2
Unidecode==0.4.21
uritools==2.0.0
urllib3==1.21.1
urlparse2==1.1.1
urlparse3==1.1
uWSGI==2.0.15
wcwidth==0.1.7
Werkzeug==0.12.2
